JP-1218 Move nightly association tests to regtest area, update test names
