+++ jwst/regtest/test_associations_generate.py
