+++ jwst/regtest/test_associations_io.py
