+++ jwst/regtest/test_associations_level3_product_names.py
