+++ jwst/regtest/test_associations_main.py
