+++ setup.cfg
@@ -28,7 +28,7 @@ ignore = E203, W503, W504, W605
 
 [tool:pytest]
 minversion = 3.6
-norecursedirs = docs/_build jwst/timeconversion jwst/extern scripts jwst/tests_nightly/general/[nmf]*
+norecursedirs = docs/_build jwst/timeconversion jwst/extern scripts jwst/tests_nightly
 asdf_schema_tests_enabled = true
 asdf_schema_root = jwst/transforms/schemas jwst/datamodels/schemas
 junit_family = xunit2
